### 2026-01-29 c2a25e0
feat: some uniconstruct scripters can now return list of model syntax, dmacs calculation made multi-factor friendly, creation of outputInvarItem to table dmacs summaries, scriptCFA() and scriptMultiCor() both deploy same functionality, update of M-CDFM vignette docs: update pkgdown
