+++ man-roxygen/fileName.R
@@ -1,0 +1,6 @@
+#' @param fileName A character string specifying a desired base name for the
+#'  output file. The default is `NULL`. The specified name will be automatically
+#'  appended with the appropriate file extension (e.g., `.txt` for `lavaan`
+#'  scripts, `.rtf` for tabular output when `gtTab = TRUE`, or other extensions
+#'  as appropriate for the output type). If a file with the same name already
+#'  exists in the user's chosen directory, it will be overwritten.
