+++ man-roxygen/scaleset.R
@@ -1,0 +1,3 @@
+#' @param scaleset Input character to specify how to set the scale of the latent
+#'  variable(s). Default is `"FF"` (fixed-factor; see Details for rationale),
+#'  but user can specify `"MV"` (Marker Variable).
