+++ R/esHelpers.R
@@ -12,50 +12,113 @@
 #'
 #' @noRd
 grouploads <- function(fit, dvn, source){
-  if(source == "1"){
-    loads <- lavaan::parameterEstimates(fit) %>%
-      dplyr::filter(.data$op == "=~" & .data$rhs %in% {{dvn}}[["p1xvarnames"]]) %>%
-      dplyr::select(.data$est)
-  }else if(source == "2"){
-    loads <- lavaan::parameterEstimates(fit) %>%
-      dplyr::filter(.data$op == "=~" & .data$rhs %in% {{dvn}}[["p2xvarnames"]]) %>%
-      dplyr::select(.data$est)
+  # Check if p1xvarnames is a list (multiple LVs) or character vector (single LV)
+  if (is.list(dvn[["p1xvarnames"]]) && !is.data.frame(dvn[["p1xvarnames"]])) {
+    # Multiple LVs case: return a list of data frames, one per LV
+    loads_list <- list()
+    if(source == "1"){
+      for (lv_name in names(dvn[["p1xvarnames"]])) {
+        loads_list[[lv_name]] <- lavaan::parameterEstimates(fit) %>%
+          dplyr::filter(.data$op == "=~" & .data$rhs %in% dvn[["p1xvarnames"]][[lv_name]]) %>%
+          dplyr::select(.data$est)
+      }
+    }else if(source == "2"){
+      for (lv_name in names(dvn[["p2xvarnames"]])) {
+        loads_list[[lv_name]] <- lavaan::parameterEstimates(fit) %>%
+          dplyr::filter(.data$op == "=~" & .data$rhs %in% dvn[["p2xvarnames"]][[lv_name]]) %>%
+          dplyr::select(.data$est)
+      }
+    }
+    return(loads_list)
+  } else {
+    # Single LV case: use original logic
+    if(source == "1"){
+      loads <- lavaan::parameterEstimates(fit) %>%
+        dplyr::filter(.data$op == "=~" & .data$rhs %in% {{dvn}}[["p1xvarnames"]]) %>%
+        dplyr::select(.data$est)
+    }else if(source == "2"){
+      loads <- lavaan::parameterEstimates(fit) %>%
+        dplyr::filter(.data$op == "=~" & .data$rhs %in% {{dvn}}[["p2xvarnames"]]) %>%
+        dplyr::select(.data$est)
+    }
+    return(loads)
   }
-  return(loads)
 }
 
 #' @rdname esHelpers
 #' @noRd
 
 groupints <- function(fit, dvn,source){
-  if(source == "1"){
-    ints <- lavaan::parameterEstimates(fit) %>%
-      dplyr::filter(.data$op == "~1" & .data$lhs %in% {{dvn}}[["p1xvarnames"]]) %>%
-      dplyr::select(.data$est)
-  }else if(source == "2"){
-    ints <- lavaan::parameterEstimates(fit) %>%
-      dplyr::filter(.data$op == "~1" & .data$lhs %in% {{dvn}}[["p2xvarnames"]]) %>%
-      dplyr::select(.data$est)
+  # Check if p1xvarnames is a list (multiple LVs) or character vector (single LV)
+  if (is.list(dvn[["p1xvarnames"]]) && !is.data.frame(dvn[["p1xvarnames"]])) {
+    # Multiple LVs case: return a list of data frames, one per LV
+    ints_list <- list()
+    if(source == "1"){
+      for (lv_name in names(dvn[["p1xvarnames"]])) {
+        ints_list[[lv_name]] <- lavaan::parameterEstimates(fit) %>%
[…]
