+++ R/scriptUni.R
@@ -12,6 +12,9 @@
 #'  user can specify `"MV"` (Marker Variable).
 #' @param lvname Input character to (arbitrarily) name the latent variable in
 #' `lavaan` syntax.
+#' @param lvar Input character to specify whether the latent variable represents
+#'  "X" or "Y" in the model. Default is `"X"`. This argument controls parameter
+#'  labeling (e.g., `lx` vs `ly` for loadings, `tx` vs `ty` for intercepts).
 #' @param constr_dy_meas Input character vector detailing which measurement
 #' model parameters to constrain across dyad members.
 #' @param constr_dy_struct Input character vector detailing which structural
@@ -105,14 +108,16 @@
 #' )
 #'
 scriptUni <- function(
-    dvn,
-    scaleset = "FF",
-    lvname = "X",
-    constr_dy_meas = c("loadings", "intercepts", "residuals"),
-    constr_dy_struct = "none", # Users do not need to modify `constr_dy_struct` when using `scriptUni()`.
-    writeTo = NULL,
-    fileName = NULL,
-    outputType = "lavaan script") {
+  dvn,
+  scaleset = "FF",
+  lvname = "X",
+  lvar = "X",
+  constr_dy_meas = c("loadings", "intercepts", "residuals"),
+  constr_dy_struct = "none", # Users do not need to modify `constr_dy_struct` when using `scriptUni()`.
+  writeTo = NULL,
+  fileName = NULL,
+  outputType = "lavaan script"
+) {
   # Input validation
   # Validate dvn argument
   if (missing(dvn) || is.null(dvn)) {
@@ -121,13 +126,29 @@ scriptUni <- function(
   if (!is.list(dvn)) {
     stop("The `dvn` argument must be a list object.")
   }
-  if (length(dvn) != 6) {
-    stop("You must supply a dvn object containing information for only X [i.e., your target LV]")
-  }
 
   # Validate lvname argument
   if (!is.character(lvname)) {
     stop("The `lvname` argument must be a character string.")
+  }
+
+  # Validate lvar argument
+  if (!lvar %in% c("X", "Y")) {
+    stop("lvar must be either 'X' or 'Y'")
+  }
+
+  # Validate dvn has required elements and only contains information for the
+  # requested latent variable family (X or Y). For historical reasons,
+  # scriptUni() expects a dvn with *only* X (or only Y) information, in a
+  # 6-element list: p1[var]varnames, p2[var]varnames, [var]indper, dist1,
+  # dist2, indnum.
+  if (lvar == "X") {
+    required_elements <- c("p1xvarnames", "p2xvarnames", "xindper", "dist1", "dist2", "indnum")
+  } else { # lvar == "Y"
+    required_elements <- c("p1yvarnames", "p2yvarnames", "yindper", "dist1", "dist2", "indnum")
+  }
+  if (!all(required_elements %in% names(dvn)) || length(dvn) != length(required_elements)) {
+    stop("You must supply a dvn object containing information for only X")
   }
 
   # check for valid inputs
@@ -156,7 +177,7 @@ scriptUni <- function(
     if (any(constr_dy_meas == "loadings")) {
       xloadsg <- loads(
         dvn,
-        lvar = "X",
+        lvar = lvar,
         lvname,
         partner = "g",
         type = "equated"
@@ -164,7 +185,7 @@ scriptUni <- function(
     } else {
       xloadsg <- loads(
         dvn,
-        lvar = "X",
+        lvar = lvar,
         lvname,
         partner = "g",
         type = "free"
@@ -175,26 +196,26 @@ scriptUni <- function(
     if (any(constr_dy_meas == "intercepts")) {
       xints1 <- intercepts(
         dvn,
-        lvar = "X",
+        lvar = lvar,
[…]
