+++ articles/M-CDFM.md
@@ -4,15 +4,17 @@
 
 This tutorial is a supplemental material from the following article:
 
-Sakaluk, J. K., & Camanto, O. J. (2025). *Dyadic Data Analysis via
-Structural Equation Modeling with Latent Variables: A Tutorial with the
-dySEM package for R*.
+Prine-Munroe, M., Sakaluk, J. K., Camanto, O., & Quinn-Nilas, C. (2025).
+*Evaluating Multi-Factor Dyadic Invariance in Couples’ Relationship
+Satisfaction Using dySEM*.
 
 This article is intended to serve as the primary citation of record for
-the `dySEM` package’s functionality for the techniques described in this
-tutorial. **If this tutorial has informed your modeling strategy
-(including but not limited to your use of `dySEM`), please cite this
-article**.
+the `dySEM` package’s functionality for fitting and interpreting the the
+Multiple Correlated Dyadic Factors Model with the help of the
+[`scriptCFA()`](https://jsakaluk.github.io/dySEM/reference/scriptCFA.md)
+function (and others) in `dySEM`. **If this tutorial has informed your
+modeling strategy (including but not limited to your use of `dySEM`),
+please cite this article**.
 
 The citation of research software aiding in analyses—like the use of
 `dySEM`—is a required practice, according to the Journal Article
@@ -80,41 +82,41 @@ library(lavaan) #for fitting dyadic SEMs
 
 For this exemplar, we use two built-in datasets from `dySEM` (focusing
 more on one). Of primary interest, the
-[`imsM`](https://jsakaluk.github.io/dySEM/reference/imsM.html) dataset
-corresponds to data from the “global” items from 282 mixed-sex couples
-responding to the full Rusbult et al. (1998) Investment Model Scale
-(IMS) (including 5 items each for satisfaction, quality of alternatives,
-investment, and commitment). More information about this data set can be
-found in [Sakaluk et al. (2021)](https://doi.org/10.1111/pere.12341).
+[`pnrqM`](https://jsakaluk.github.io/dySEM/reference/pnrqM.html) dataset
+corresponds to data from a short-form of the Positive-Negative
+Relationship Quality Scale (PNRQ; Rogge et al., 2017) from 219
+(M)ixed-sex couples (including 4 items each for positive and negative).
+More information about this data set can be found in Prine et al. (under
+review).
 
-All variables in this data frame follow a [“sip” naming pattern](NA):
+All variables in this data frame follow a [“sip” naming pattern](NA),
+with a delimiting “\_” between the item number and distinguishing
+partner character:
 
 ``` r
 
-imsM_dat <- imsM
+pnrqM_dat <- pnrqM
 
-imsM_dat |> 
+pnrqM_dat |> 
   as_tibble()
-#> # A tibble: 282 × 40
-#>    sat.g1_f sat.g2_f sat.g3_f sat.g4_f sat.g5_f qalt.g1_f qalt.g2_f qalt.g3_f
-#>       <int>    <int>    <int>    <int>    <int>     <int>     <int>     <int>
-#>  1        2        1        1        2        1         3         1         5
-#>  2        9        9        9        9        9         5         1         1
[…]
