+++ reference/scriptUni.md
@@ -12,10 +12,12 @@ scriptUni(
   dvn,
   scaleset = "FF",
   lvname = "X",
+  lvar = "X",
   constr_dy_meas = c("loadings", "intercepts", "residuals"),
   constr_dy_struct = "none",
   writeTo = NULL,
-  fileName = NULL
+  fileName = NULL,
+  outputType = "lavaan script"
 )
 ```
 
@@ -36,6 +38,13 @@ scriptUni(
 
   Input character to (arbitrarily) name the latent variable in `lavaan`
   syntax.
+
+- lvar:
+
+  Input character to specify whether the latent variable represents "X"
+  or "Y" in the model. Default is `"X"`. This argument controls
+  parameter labeling (e.g., `lx` vs `ly` for loadings, `tx` vs `ty` for
+  intercepts).
 
 - constr_dy_meas:
 
@@ -72,10 +81,19 @@ scriptUni(
   the same name already exists in the user's chosen directory, it will
   be overwritten.
 
+- outputType:
+
+  Character string specifying the type of output to return. Options are
+  `"lavaan script"` (default) to return a character object of `lavaan`
+  syntax that can be passed immediately to `lavaan` functions, or
+  `"syntax components"` to return a structured list of model components.
+
 ## Value
 
 Character object of `lavaan` script that can be passed immediately to
-`lavaan` functions.
+`lavaan` functions (when `outputType = "lavaan script"`), or a
+structured list of model components (when
+`outputType = "syntax components"`).
 
 ## Details
 
