### 2026-02-13 #9759 @samhforbes
Hi @keahtandon @katschueler @ssriv would you have any interest in reviewing this interesting package for us at JOSS?
