### 2026-02-19 0725dce
docs: update M-CDFM vignette with path diagrams and explanation
