+++ vignettes/articles/M-CDFM.Rmd
@@ -15,7 +15,7 @@ This tutorial is a supplemental material from the following article:
 
 Prine-Munroe, M., Sakaluk, J. K., Camanto, O., & Quinn-Nilas, C. (2025). *Evaluating Multi-Factor Dyadic Invariance in Couples’ Relationship Satisfaction Using dySEM*. 
 
-This article is intended to serve as the primary citation of record for the `dySEM` package's functionality for fitting and interpreting the the Multiple Correlated Dyadic Factors Model with the help of the `scriptCFA()` function (and others) in `dySEM`. **If this tutorial has informed your modeling strategy (including but not limited to your use of `dySEM`), please cite this article**. 
+This article is intended to serve as the primary citation of record for the `dySEM` package's functionality for fitting and interpreting the the Multiple Correlated Dyadic Factors Model with the help of the `scriptMultiCor()`/`scriptCFA()` function (and others) in `dySEM`. **If this tutorial has informed your modeling strategy (including but not limited to your use of `dySEM`), please cite this article**. 
 
 The citation of research software aiding in analyses---like the use of `dySEM`---is a required practice, according to the Journal Article Reporting Standards (JARS) for Quantitative Research in Psychology ([Appelbaum et al., 2018](https://psycnet.apa.org/fulltext/2018-00750-002.html)). 
 
@@ -392,3 +392,59 @@ outputInvarItem(dvn = dvnPNRQ, fit = pnrq.partial.mod,
 
 
 ```
+
+## Quality Assurance
+
+To paraphrase the introduction to the TV show [Review](https://www.imdb.com/title/tt2141913/), the astute reader might find themselves wondering: 
+
+> `dySEM`: it’s all we have for automating dyadic SEM…but is it *any good*? 
+
+That is, how can a scrutinous reader have faith that the software actually functions, as intended, and generates the correct model syntax? 
+
+### Why have confidence in `dySEM`?
+
+All software, including ours, will inevitably contain errors. But we see three reasons for users to have confidence in `dySEM`’s core functionality. 
+
+1. Its codebase is completely open-source, meaning its entire codebase is available for interested users to inspect, test, and/or build on themselves (e.g., the code for the M-CDFA scripting is available [here](https://github.com/jsakaluk/dySEM/blob/master/R/scriptCFA.R)). 
+
+2. `dySEM` has been accepted to [CRAN](https://cran.r-project.org/web/packages/dySEM/index.html)  should give users confidence, because  [CRAN demands a strict set of checks](https://r-pkgs.org/release.html) that need to be passed in order to ensure cross-platform compatibility and proper deployment. And finally, 
+
+3. `dySEM` is backed by more than 500 [self-verifying “unit tests”](https://r-pkgs.org/testing-basics.html) that ensure its functionality executes as intended, including [31 dedicated to testing the M-CDFM functionality](https://github.com/jsakaluk/dySEM/blob/master/tests/testthat/test-scriptCFA.R). 
+
+### Key human-written and human-verified unit tests
+
+Among these unit tests, many of the most important ones have been human-written and human-verified to provide correct model specifications. Namely, for each of the four M-CDFM's (with our pnrqM data) corresponding to the four levels of invariance (configural, loading, intercept, residual), we have: 
+
+- calculated the appropriate number of parameters to be estimated, 
+- calculated the appropriate number of degrees of freedom, and
+- ensured that these values do not change across `scalset = "FF"` or `scaleset = "MV"` options.
+
[…]
