+++ reference/outputInvarCompTab.md
@@ -141,7 +141,7 @@ outputInvarCompTab(mods,
   parsimonyFirst = FALSE,
   gtTab = TRUE, writeTo = tempdir(), fileName = "dCFA_Invar_Standard"
 )
-#> Output stored in: /tmp/Rtmpu2uJhH/dCFA_Invar_Standard.rtf
+#> Output stored in: /tmp/Rtmpujbx1D/dCFA_Invar_Standard.rtf
 
 
   
@@ -302,7 +302,7 @@ sat.loading.fit, sat.intercept.fit, sat.residual.fit)
 outputInvarCompTab(mods, parsimonyFirst = TRUE, gtTab = TRUE, writeTo =
 [tempdir](https://rdrr.io/r/base/tempfile.html)(), fileName =
 "dCFA_Invar_Reverse" ) \#\> Output stored in:
-/tmp/Rtmpu2uJhH/dCFA_Invar_Reverse.rtf
+/tmp/Rtmpujbx1D/dCFA_Invar_Reverse.rtf
 
 | mod        |  chisq |  df | pvalue |      aic |      bic | rmsea |   cfi | chisq_diff | df_diff | p_diff | aic_diff | bic_diff | rmsea_diff | cfi_diff |
 |:-----------|-------:|----:|-------:|---------:|---------:|------:|------:|-----------:|--------:|-------:|---------:|---------:|-----------:|---------:|
