+++ reference/outputUniConstructComp.md
@@ -230,7 +230,7 @@ outputUniConstructComp(
   writeTo = tempdir(),
   fileName = "uni-construct-dyad-models"
 )
-#> Output stored in: /tmp/Rtmpu2uJhH/uni-construct-dyad-models.rtf
+#> Output stored in: /tmp/Rtmpujbx1D/uni-construct-dyad-models.rtf
 
 
   
