+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/L-APIM: L-APIM.html
   articles/M-CDFM: M-CDFM.html
   varnames: varnames.html
-last_built: 2026-01-29T15:37Z
+last_built: 2026-02-19T15:39Z
 urls:
   reference: https://jsakaluk.github.io/dySEM/reference
   article: https://jsakaluk.github.io/dySEM/articles
