### 2026-02-19 4f0d8e2
feat: add scriptTwoCross() for boutique two-construct models; add outputType argument for other uni-construct scripters; add marker-variable scale-setting for loads() when g, and lvar = Y tests: add tests for scriptTwoCross docs: update pkgdown
