+++ tests/testthat/test-scriptTwoCross.R
@@ -1,0 +1,153 @@
+test_that("scriptTwoCross requires dvn with both X and Y", {
+  dvn_x_only <- scrapeVarCross(
+    dat = commitmentQ, x_order = "spi", x_stem = "sat.g",
+    x_delim1 = ".", x_delim2 = "_",
+    distinguish_1 = "1", distinguish_2 = "2", verbose = FALSE
+  )
+  expect_error(
+    scriptTwoCross(dvn_x_only, lvxname = "Sat", lvyname = "Com"),
+    "both X and Y"
+  )
+})
+
+test_that("scriptTwoCross produces valid syntax for cor/uni", {
+  dvn <- scrapeVarCross(
+    dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".", x_delim2 = "_",
+    distinguish_1 = "1", distinguish_2 = "2",
+    y_order = "spi", y_stem = "com", y_delim1 = ".", y_delim2 = "_",
+    verbose = FALSE
+  )
+  script <- scriptTwoCross(dvn, lvxname = "Sat", lvyname = "Com",
+    x_model = "cor", y_model = "uni")
+  expect_type(script, "character")
+  expect_gt(nchar(script), 500)
+  expect_true(grepl("Sat1", script))
+  expect_true(grepl("ComDy", script))
+  expect_true(grepl("Latent Regressions", script))
+})
+
+test_that("scriptTwoCross produces valid syntax for all 16 model combinations", {
+  dvn <- scrapeVarCross(
+    dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".", x_delim2 = "_",
+    distinguish_1 = "1", distinguish_2 = "2",
+    y_order = "spi", y_stem = "com", y_delim1 = ".", y_delim2 = "_",
+    verbose = FALSE
+  )
+  models <- c("uni", "cor", "hier", "bifactor")
+  for (xm in models) {
+    for (ym in models) {
+      script <- scriptTwoCross(dvn, lvxname = "Sat", lvyname = "Com",
+        x_model = xm, y_model = ym)
+      expect_type(script, "character")
+      expect_true(nchar(script) > 100, label = paste("x_model =", xm, ", y_model =", ym))
+    }
+  }
+})
+
+test_that("scriptTwoCross constr_dy_xy_struct zero yields zero regression paths", {
+  dvn <- scrapeVarCross(
+    dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".", x_delim2 = "_",
+    distinguish_1 = "1", distinguish_2 = "2",
+    y_order = "spi", y_stem = "com", y_delim1 = ".", y_delim2 = "_",
+    verbose = FALSE
+  )
+  script <- scriptTwoCross(dvn, lvxname = "Sat", lvyname = "Com",
+    x_model = "cor", y_model = "uni", constr_dy_xy_struct = "zero")
+  expect_true(grepl("0\\*Sat1", script))
+  expect_true(grepl("0\\*Sat2", script))
+})
+
+test_that("scriptTwoCross rejects invalid x_model and y_model", {
+  dvn <- scrapeVarCross(
+    dat = commitmentQ, x_order = "spi", x_stem = "sat.g", x_delim1 = ".", x_delim2 = "_",
+    distinguish_1 = "1", distinguish_2 = "2",
+    y_order = "spi", y_stem = "com", y_delim1 = ".", y_delim2 = "_",
+    verbose = FALSE
+  )
[…]
