+++ R/scriptHier.R
@@ -9,14 +9,22 @@
 #' @param dvn Input dvn list from `scrapeVarCross()`
 #' @template scaleset
 #' @param lvname Input character to (arbitrarily) name the latent variable in `lavaan` syntax
+#' @param lvar Input character to specify whether the latent variable represents
+#'  "X" or "Y" in the model. Default is `"X"`. This argument controls parameter
+#'  labeling and which indicator variables are used from the dvn.
 #' @param constr_dy_meas Input character vector detailing which measurement model parameters to constrain across dyad members.
 #' @param constr_dy_struct Input character vector detailing which structural model parameters to constrain across dyad members.
 #' Default is `c("variances", "means")`(in combination with defaults for `constr_dy_meas`, an indistinguishable correlated dyadic factors model),
 #' but user can specify any combination of `"variances"` and `"means"`, or `"none"`.
 #' @template writeTo
 #' @template fileName
+#' @param outputType Character string specifying the type of output to return.
+#'  Options are `"lavaan script"` (default) to return a character object of
+#'  `lavaan` syntax that can be passed immediately to `lavaan` functions, or
+#'  `"syntax components"` to return a structured list of model components.
 #' @return Character object of `lavaan` script that can be passed immediately to
-#'  `lavaan` functions.
+#'  `lavaan` functions (when `outputType = "lavaan script"`), or a structured list
+#'  of model components (when `outputType = "syntax components"`).
 #'
 #' @details
 #' * By default, many `dySEM::` functions (including `scriptHier()`) default to
@@ -105,15 +113,22 @@ scriptHier <- function(
     dvn,
     scaleset = "FF",
     lvname = "X",
+    lvar = "X",
     constr_dy_meas = c("loadings", "intercepts", "residuals"),
     constr_dy_struct = c("variances", "means"),
     writeTo = NULL,
-    fileName = NULL
+    fileName = NULL,
+    outputType = "lavaan script"
 ){
 
   #check for valid inputs
-  if(length(dvn)!=6){
-    stop("You must supply a dvn object containing information for only X [i.e., your target LV]")
+  if (lvar == "X") {
+    required_elements <- c("p1xvarnames", "p2xvarnames", "xindper", "dist1", "dist2", "indnum")
+  } else {
+    required_elements <- c("p1yvarnames", "p2yvarnames", "yindper", "dist1", "dist2", "indnum")
+  }
+  if (!all(required_elements %in% names(dvn)) || length(dvn) != 6) {
+    stop("You must supply a dvn object containing information for only X or Y [i.e., your target LV]")
   }
 
   if(!any(constr_dy_meas %in% c("loadings", "intercepts", "residuals", "none"))){
@@ -128,16 +143,24 @@ scriptHier <- function(
     stop("scaleset must be either 'FF' (fixed-factor) or 'MV' (marker variable)")
   }
 
+  if (!outputType %in% c("lavaan script", "syntax components")) {
+    stop("outputType must be either 'lavaan script' or 'syntax components'")
+  }
+
+  if (!lvar %in% c("X", "Y")) {
+    stop("lvar must be either 'X' or 'Y'")
+  }
+
   if(scaleset == "FF"){
 
     #loadings
[…]

[…]
