### 2026-02-19 dda3f7e
tests: add two lightweight tests ensuring similarity across chi2 and df for scriptTwoCross() and scriptAPIM()
