### 2026-02-20 5237fb9
feat: add multigroup functionality to scraper, helpers, and uniconstruct scripters (#155). Works for cor and uni
