+++ R/scriptUni.R
@@ -23,6 +23,8 @@
 #'  constrain across the modeled dyad members. For consistency with other
 #'  scripter functions, `constr_dy_struct` is included as an argument, but
 #'  defaults to `"none"`.
+#' @param constr_group_meas Optional character vector for cross-group measurement constraints (e.g. \code{"loadings"}, \code{"intercepts"}, \code{"residuals"}, \code{"none"}). When \code{NULL} (default), single-group model. When non-NULL, requires \code{dvn} from \code{scrapeVarCross(..., group = "varname")}.
+#' @param constr_group_struct Optional character vector for cross-group structural constraints (e.g. \code{"variances"}, \code{"means"}, \code{"none"}). When \code{NULL} (default), single-group model.
 #' @template writeTo
 #' @template fileName
 #' @param outputType Character string specifying the type of output to return.
@@ -112,6 +114,8 @@ scriptUni <- function(
   lvar = "X",
   constr_dy_meas = c("loadings", "intercepts", "residuals"),
   constr_dy_struct = "none", # Users do not need to modify `constr_dy_struct` when using `scriptUni()`.
+  constr_group_meas = NULL,
+  constr_group_struct = NULL,
   writeTo = NULL,
   fileName = NULL,
   outputType = "lavaan script"
@@ -145,7 +149,7 @@ scriptUni <- function(
   } else { # lvar == "Y"
     required_elements <- c("p1yvarnames", "p2yvarnames", "yindper", "dist1", "dist2", "indnum")
   }
-  if (!all(required_elements %in% names(dvn)) || length(dvn) != length(required_elements)) {
+  if (!all(required_elements %in% names(dvn))) {
     stop("You must supply a dvn object containing information for only X")
   }
 
@@ -155,18 +159,64 @@ scriptUni <- function(
     stop("scaleset must be either 'FF' (fixed-factor) or 'MV' (marker variable)")
   }
 
-  if (!any(constr_dy_meas %in% c("loadings", "intercepts", "residuals", "none"))) {
-    stop("constr_dy_meas must be a character vector containing any combination of 'loadings', 'intercepts', 'residuals', or 'none'")
+  valid_dy_meas <- c("loadings", "intercepts", "residuals", "none")
+  invalid_dy_meas <- setdiff(constr_dy_meas, valid_dy_meas)
+  if (length(invalid_dy_meas) > 0) {
+    stop("constr_dy_meas contains invalid value(s): ", paste(sQuote(invalid_dy_meas), collapse = ", "),
+         ". Valid options: ", paste(sQuote(valid_dy_meas), collapse = ", "))
   }
 
-  if (!any(constr_dy_struct %in% c("none"))) {
-    stop("constr_dy_struct is not applicable to `scriptUni()`.
-          Please leave it as the default value: 'none'.")
+  valid_dy_struct <- c("none")
+  invalid_dy_struct <- setdiff(constr_dy_struct, valid_dy_struct)
+  if (length(invalid_dy_struct) > 0) {
+    stop("constr_dy_struct contains invalid value(s): ", paste(sQuote(invalid_dy_struct), collapse = ", "),
+         ". Valid options: ", paste(sQuote(valid_dy_struct), collapse = ", "))
   }
 
   # Validate outputType argument
   if (!outputType %in% c("lavaan script", "syntax components")) {
     stop("outputType must be either 'lavaan script' or 'syntax components'")
+  }
+
+  # Multi-group: validate constr_group_* and derive group_n / constr flags
+  group_n <- NULL
+  constr_group_loadings <- FALSE
+  constr_group_intercepts <- FALSE
+  constr_group_residuals <- FALSE
[…]

[…]
