### 2026-02-21 #9759 @samhforbes
Hi @smasongarrison & @lionel68 would you have any interest in reviewing this interesting package for us at JOSS?
