### 2026-02-21 #9759 @smasongarrison
I'd be happy to review. From a first skim, I'd be really interested in diving deeper into this!
