### 2026-02-23 #9759 @keahtandon
Yes, I am interested in reviewing this if you still need me. This will be my first review, so I may need guidance.

Thanks,
Keah

From: Sam Forbes ***@***.***>
Sent: Friday, February 13, 2026 12:04 PM
Cc: Tandon, Keah ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: dySEM: An R Package for Dyadic Structural Equation Modeling with Latent Variables (Issue #9759)

[https://avatars.githubusercontent.com/u/18288098?s=20&v=4]samhforbes left a comment (openjournals/joss-reviews#9759)<https://github.com/openjournals/joss-reviews/issues/9759#issuecomment-3898278281>

Hi @keahtandon<https://github.com/keahtandon> @katschueler<https://github.com/katschueler> @ssriv<https://github.com/ssriv> would you have any interest in reviewing this interesting package for us at JOSS?

-
You are receiving this because you were mentioned.Message ID: ***@***.******@***.***>>
