### 2026-02-23 39a0697
fix: corrected placement of higher-order loading statement; models involving Hier only predict/are predicted by higher-order; models involving bifactor as predictor only use g (except for y-bifactor); uni and cor can predictor g and specific; bifactor g predicts g, and specifics predict specifics
