### 2026-02-24 35fbf37
feat: add scripting functionality (helpers and scripters) for simulation populations models (and documentation and tests)
