+++ tests/testthat/test-scriptCor.R
@@ -654,7 +654,7 @@ test_that("scriptCor rejects invalid constr_dy_meas values", {
   )
   expect_error(
     scriptCor(dvn = dvn, lvname = "Sat", constr_dy_meas = "invalid_option"),
-    "constr_dy_meas must be a character vector containing any combination of 'loadings', 'intercepts', 'residuals', or 'none'"
+    "constr_dy_meas contains invalid value\\(s\\)|Valid options"
   )
 })
 
@@ -666,7 +666,7 @@ test_that("scriptCor rejects invalid constr_dy_struct values", {
   )
   expect_error(
     scriptCor(dvn = dvn, lvname = "Sat", constr_dy_struct = "invalid_option"),
-    "constr_dy_struct must be a character vector containing any combination of 'variances', 'means', or 'none'"
+    "constr_dy_struct contains invalid value\\(s\\)|Valid options"
   )
 })
 
