### 2026-02-25 #10115 @editorialbot
[REVIEW]: dySEM: An R Package for Dyadic Structural Equation Modeling with Latent Variables
**Submitting author:** @jsakaluk (<a href="http://orcid.org/0000-0002-2515-9822">John Kitchener Sakaluk</a>)
**Repository:** https://github.com/jsakaluk/dySEM
**Branch with paper.md** (empty if default branch):
**Version:** 1.4.1
