### 2026-02-25 27e10b8
feat: add coresids_zero and coresids_equal to constraint options; add new options to coresids in scriptHelper
