+++ R/scriptBifac.R
@@ -178,11 +178,14 @@ scriptBifac <- function(
     stop("scaleset must be either 'FF' (fixed-factor) or 'MV' (marker variable)")
   }
 
-  valid_dy_meas <- c("loadings", "loadings_source", "intercepts", "residuals", "none")
+  valid_dy_meas <- c("loadings", "loadings_source", "intercepts", "residuals", "coresids_zero", "coresids_equal", "none")
   invalid_dy_meas <- setdiff(constr_dy_meas, valid_dy_meas)
   if (length(invalid_dy_meas) > 0) {
     stop("constr_dy_meas contains invalid value(s): ", paste(sQuote(invalid_dy_meas), collapse = ", "),
          ". Valid options: ", paste(sQuote(valid_dy_meas), collapse = ", "))
+  }
+  if (any(constr_dy_meas == "coresids_zero") && any(constr_dy_meas == "coresids_equal")) {
+    stop("constr_dy_meas cannot contain both 'coresids_zero' and 'coresids_equal'.")
   }
 
   valid_dy_struct <- c("variances", "means", "none")
@@ -403,10 +406,11 @@ scriptBifac <- function(
     }
 
     # correlated residuals
+    coresids_type <- if (any(constr_dy_meas == "coresids_zero")) "zero" else if (any(constr_dy_meas == "coresids_equal")) "equated" else "free"
     xcoresids <- coresids(
       dvn,
       lvar = lvar,
-      type = "free",
+      type = coresids_type,
       group_n = group_n,
       constr_group_residual_covariances = constr_group_residual_covariances
     )
@@ -813,7 +817,8 @@ scriptBifac <- function(
 
 
     # correlated residuals for X
-    xcoresids <- coresids(dvn, lvar = lvar, type = "free", group_n = group_n, constr_group_residual_covariances = constr_group_residual_covariances)
+    coresids_type <- if (any(constr_dy_meas == "coresids_zero")) "zero" else if (any(constr_dy_meas == "coresids_equal")) "equated" else "free"
+    xcoresids <- coresids(dvn, lvar = lvar, type = coresids_type, group_n = group_n, constr_group_residual_covariances = constr_group_residual_covariances)
 
 
     # latent variances

[…]
