### 2026-03-02 a7f930f
docs: update to 2.0 for DESCRIPTION, NAMESPACE, README, and pkgdown site
