+++ man/scriptPopHelpers.Rd
@@ -1,46 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/scriptPopHelpers.R
-\name{scriptPopHelpers}
-\alias{scriptPopHelpers}
-\alias{popLoadings}
-\alias{popIntercepts}
-\alias{popResiduals}
-\alias{popCoresids}
-\alias{popLatentVars}
-\alias{popLatentCovars}
-\alias{popLatentMeans}
-\title{Simulation Population Model Helpers}
-\usage{
-popLoadings(dvn, lvname, partner, values, lvar = "X")
-
-popIntercepts(dvn, partner, values, lvar = "X")
-
-popResiduals(dvn, partner, values, lvar = "X")
-
-popCoresids(dvn, values, lvar = "X")
-
-popLatentVars(dvn, lvname, partner, value, lvar = "X")
-
-popLatentCovars(dvn, lvname, value, lvar = "X")
-
-popLatentMeans(dvn, lvname, partner, value, lvar = "X")
-}
-\arguments{
-\item{dvn}{Input dvn list from \code{build_dvn()} or \code{scrapeVarCross()}}
-
-\item{lvname}{Character name for the latent variable in lavaan syntax}
-
-\item{partner}{Character: \code{"1"} or \code{"2"} for partner-specific, or \code{"g"} for unidimensional (both partners)}
-
-\item{values}{Numeric vector of parameter values. Length must match number of indicators.}
-
-\item{lvar}{Character \code{"X"} or \code{"Y"} for latent variable family. Default \code{"X"}.}
-
-\item{value}{Numeric; latent mean value}
-}
-\description{
-Helper functions for writing lavaan syntax with fixed numeric values
-for population models used in Monte Carlo simulations. All helpers require
-explicit numeric values (no heuristics).
-}
-\concept{population-scripting helpers}
