+++ NAMESPACE
@@ -14,13 +14,6 @@ export(outputModel)
 export(outputParamFig)
 export(outputParamTab)
 export(outputUniConstructComp)
-export(popCoresids)
-export(popIntercepts)
-export(popLatentCovars)
-export(popLatentMeans)
-export(popLatentVars)
-export(popLoadings)
-export(popResiduals)
 export(scrapeVarCross)
 export(scriptAPIM)
 export(scriptBiDy)
