+++ reference/scriptTwoCross.md
@@ -1,0 +1,174 @@
+# A Function That Writes, Saves, and Exports Syntax for Fitting Boutique Two-Construct Dyadic SEM Models
+
+This function takes the outputted object from
+[`scrapeVarCross()`](https://jsakaluk.github.io/dySEM/reference/scrapeVarCross.md)
+with both X and Y indicators and automatically writes, returns, and
+exports (`.txt`) `lavaan` syntax for specifying a two-construct model
+where the measurement structure of X and Y can differ (e.g., X as
+correlated factors, Y as unidimensional).
+
+## Usage
+
+``` r
+scriptTwoCross(
+  dvn,
+  x_scaleset = "FF",
+  y_scaleset = "FF",
+  lvxname = "X",
+  lvyname = "Y",
+  x_model = "cor",
+  y_model = "cor",
+  constr_dy_x_meas = c("loadings", "intercepts", "residuals"),
+  constr_dy_x_struct = c("variances", "means"),
+  constr_dy_y_meas = c("loadings", "intercepts", "residuals"),
+  constr_dy_y_struct = c("variances", "means"),
+  constr_dy_xy_struct = "free",
+  includeMeanStruct = FALSE,
+  writeTo = NULL,
+  fileName = NULL
+)
+```
+
+## Arguments
+
+- dvn:
+
+  Input dvn list from
+  [`scrapeVarCross()`](https://jsakaluk.github.io/dySEM/reference/scrapeVarCross.md)
+  containing information for both X and Y (i.e., must include both
+  x_stem and y_stem arguments when calling scrapeVarCross).
+
+- x_scaleset:
+
+  Character string specifying how to set the scale of latent X. Options
+  are `"FF"` (fixed-factor) or `"MV"` (marker variable). Default is
+  `"FF"`.
+
+- y_scaleset:
+
+  Character string specifying how to set the scale of latent Y. Options
+  are `"FF"` (fixed-factor) or `"MV"` (marker variable). Default is
+  `"FF"`.
+
+- lvxname:
+
+  Input character to (arbitrarily) name latent variable X in `lavaan`
+  syntax.
+
+- lvyname:
+
+  Input character to (arbitrarily) name latent variable Y in `lavaan`
+  syntax.
+
+- x_model:
+
+  Character string specifying the measurement model for X. Options are
+  `"uni"` (unidimensional), `"cor"` (correlated factors), `"hier"`
+  (hierarchical), or `"bifactor"`.
+
+- y_model:
+
+  Character string specifying the measurement model for Y. Same options
+  as `x_model`.
+
+- constr_dy_x_meas:
+
+  Input character vector detailing which measurement model parameters to
+  constrain across dyad members for latent X.
+
+- constr_dy_x_struct:
+
+  Input character vector detailing which structural model parameters to
+  constrain across dyad members for latent X.
+
[…]
