### 2026-03-04 #10115 @keahtandon
@samhforbes When I'm reading the article, in the State of the Field section, there appears to be '??' in place of 'Table 1' in text. Is that something that I should ask the author about in the repository, or is there a better place to mention that?
