### 2026-03-24 #10115 @jsakaluk
@samhforbes may correct me, but I'm under the impression that the '??' is autopopulating in a fashion that is (as hosted on GitHub) correct, but if/when accepted to JOSS, will populate with the appropriate table number.

(and thanks to you both, for your reviewing efforts so far!)
