### 2026-03-25 #10115 @samhforbes
Yes I believe so, but we can keep an eye on it when we move towards final compilation!
