### 2026-04-05 #10115 @smasongarrison
@jsakaluk I really love your vignette on variable naming, especially the infographics!
