### 2026-04-05 52fdd07
Update paper.md
attempting to resolve url link, also caught a typo
