### 2026-04-05 623f613
Update dySEM.Rmd
fixed so that example can run
