+++ man/DRES.Rd
@@ -43,7 +43,7 @@ data(DRES)
 \description{
 A dataset containing 9 observed indicators of relationship quality (PRQC)
 and 5 observed indicators of sexual satisfaction from 121 couples
-in the DRES (Daily Relationship Experiences Study; Raposo, Impett, & Muise, in press)
+in the DRES (Daily Relationship Experiences Study; Raposo, Impett, & Muise, 2020)
 }
 \references{
 Raposo, S., Impett, E. A., & Muise, A. (2020). Avoidantly Attached Individuals Are More Exchange‐Oriented and Less Communal in the Bedroom. \emph{Archives of Sexual Behavior}, \emph{49}, 2863–2881. https://doi.org/10.1007/s10508-020-01813-9
