### 2026-04-05 #10115 @smasongarrison
@jsakaluk I noticed that a couple of the dataset references are a bit out of date. I added the ones that were easy to fix to my [pull](https://github.com/jsakaluk/dySEM/pull/156) request. The only one I couldn't easily find was for `pnrqM`. (The pull request just has some tiny fixes and should help with the weird url wrapping on line 48.) Nice job on this!
