### 2026-04-07 522055e
feat: new GUI and APIM scripter for simulation fix: orthogonal argumentation for some scripters docs: update pkgdown, DESCRIPTION, NAMESPACE, and NEWs
