+++ paper/paper.bib
@@ -1,63 +1,55 @@
 
 @book{littleLongitudinalStructuralEquation2013,
-	location = {New York, {NY}},
+	address = {New York, NY},
 	title = {Longitudinal structural equation modeling},
 	publisher = {Guilford Press},
 	author = {Little, Todd D.},
-	date = {2013},
+	year = {2013},
 }
 
 @book{kennyDyadicDataAnalysis2006,
-	location = {New York, {NY}},
+	address = {New York, NY},
 	title = {Dyadic data analysis},
 	publisher = {Guilford Press},
 	author = {Kenny, David A. and Kashy, Deborah A. and Cook, William L.},
-	date = {2006},
+	year = {2006},
 }
 
 @article{mccoachDealingDependencePart2010,
-	title = {Dealing With Dependence (Part I): Understanding the Effects of Clustered Data},
+	title = {Dealing {With} {Dependence} ({Part} {I}): {Understanding} the {Effects} of {Clustered} {Data}},
 	volume = {54},
 	issn = {0016-9862, 1934-9041},
+	shorttitle = {Dealing {With} {Dependence} ({Part} {I})},
 	url = {http://journals.sagepub.com/doi/10.1177/0016986210363076},
 	doi = {10.1177/0016986210363076},
-	shorttitle = {Dealing With Dependence (Part I)},
-	abstract = {This article provides a conceptual introduction to the issues surrounding the analysis of clustered (nested) data. We define the intraclass correlation coefficient ({ICC}) and the design effect, and we explain their effect on the standard error. When the {ICC} is greater than 0, then the design effect is greater than 1. In such a scenario, the standard error produced under the assumption of independence is underestimated. This increases the Type I error rate. We provide a short illustration of the effect of non-independence on the standard error.We show that after accounting for the design effect, our decision about the statistical significance of the test statistic changes.When we fail to account for the clustered nature of the data, we conclude that the difference between the two groups is statistically significant. However, once we adjust the standard error for the design effect, the difference is no longer statistically significant.},
-	pages = {152--155},
+	abstract = {This article provides a conceptual introduction to the issues surrounding the analysis of clustered (nested) data. We define the intraclass correlation coefficient (ICC) and the design effect, and we explain their effect on the standard error. When the ICC is greater than 0, then the design effect is greater than 1. In such a scenario, the standard error produced under the assumption of independence is underestimated. This increases the Type I error rate. We provide a short illustration of the effect of non-independence on the standard error.We show that after accounting for the design effect, our decision about the statistical significance of the test statistic changes.When we fail to account for the clustered nature of the data, we conclude that the difference between the two groups is statistically significant. However, once we adjust the standard error for the design effect, the difference is no longer statistically significant.},
+	language = {en},
 	number = {2},
-	journaltitle = {Gifted Child Quarterly},
-	author = {{McCoach}, D. Betsy and Adelson, Jill L.},
 	urldate = {2018-06-04},
-	date = {2010-04},
-	langid = {english},
+	journal = {Gifted Child Quarterly},
+	author = {McCoach, D. Betsy and Adelson, Jill L.},
+	month = apr,
+	year = {2010},
+	pages = {152--155},
 	file = {McCoach and Adelson - 2010 - Dealing With Dependence (Part I) Understanding th.pdf:/Users/johnsakaluk/Zotero/storage/TQIA8QNH/McCoach and Adelson - 2010 - Dealing With Dependence (Part I) Understanding th.pdf:application/pdf},
 }
 
-@book{rcoreteamLanguageEnvironmentStatistical2022,
[…]
