+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/L-APIM: L-APIM.html
   articles/M-CDFM: M-CDFM.html
   varnames: varnames.html
-last_built: 2026-04-07T20:16Z
+last_built: 2026-04-07T20:53Z
 urls:
   reference: https://jsakaluk.github.io/dySEM/reference
   article: https://jsakaluk.github.io/dySEM/articles
