### 2026-04-08 #10115 @jsakaluk
@keahtandon and @smasongarrison , thanks for your careful eyes, and time and energy on this review!

@smasongarrison : I can't take credit for the illustrations--a talented undergraduate with a knack for visuals did some excellent work on that piece! I've reviewed and merged your PR, and all the GH actions are passing. I also updated the wayward DOIs and the bot seems pleased. re: the reference for pnrqM, I gesture at the in-review paper in the documentation; I will update this if/when Merissa's (Prine) paper is published (as she and Chris Quinn-Nilas were the ones who collected/contributed the data)
