### 2026-04-08 #10115 @smasongarrison
@samhforbes I'm happy :)

@jsakaluk No worries about the Prine paper being still in review. I guess I had been hopeful that it would have escaped the review process by now. (on the positive it means that you'll be able to cite the Joss paper in that paper 🤣 )
