### 2026-04-13 #10115 @jsakaluk
@smasongarrison alas, a recent R&R, so still awhile yet until that one sees the light of day. But as you say, a good opportunity to cite the JOSS paper :P Also, unrelated to the package, but was peaking around, and that looks like a helluva data science course you've built!
