### 2026-04-15 #10115 @jsakaluk
@samhforbes, wonderful. A few clarifying questions (and some confusion, which may be a function of my mis-understanding of JOSS processes):

- does CRAN satisfy the piece around archiving the software and DOI-minting (i.e., in lieu of Zenodo and/or figshare)? One reason for asking is that the package is already on CRAN/will continue to be maintained there, it already has a DOI (10.32614/CRAN.package.dySEM), and I have some reservations about what it means for me to archive/maintain the package in multiple locations, and for users to navigate multiple DOIs for the software
- relatedly, given the required changes were more editorial, than bearing on any functionality direclty (i.e., changing an argument in the initial vignette, and correcting years in documentation for data sets), I'm wondering if the 1.4.1 release on CRAN can stand as the version to be released in the paper? Part of this request is perhaps me misunderstanding what I should have done upon submission (i.e., stopped development, and/or kept development in a separate branch and withhold merging until this review process was done), but at this point, I've pushed ahead to ver. 2.0.0, and I'm a bit worried that if I go to submit coverage of the new functionality (of which there is a considerable amount) in a paper somewhere, linking the final version of this paper (which does not mention/demo that functionality) to the release containing that functionality, might prohibit me from doing so (e.g., someone(s) might feel like I'm double-dipping). If there was some guidance of how to continue developing without mucking up this part of the JOSS submission process, my apologies for overlooking it. Any suggestions you have of how to thread the needle on this, I'm happy to consider!
