### 2026-04-16 #10115 @samhforbes
Hi @jsakaluk good questions!

- CRAN typically does not satisfy that requirement, since the DOI should include everything as it was reviewed here. That said that doesn't need maintenance. What often happens is people will put on eg Zenodo a DOI'd release that explicitly says it is the JOSS paper release.
- Then relatedly, we would just expect that the package is feature complete, so it should be the current version in that state. That doesn't preclude you though from writing more detailed methods papers elsewhere if I've understood your meaning.
