### 2026-04-22 #10115 @jsakaluk
@samhforbes
- I think my concern/confusion was that, having developed more functionality in the time that has passed, having a package version that now includes the changes made as a part of this review AND that new functionality all under one released version, would other journals balk at reviewing submissions around that new functionality. Essentially, would it look like "double dipping" if the version of origin for the new functionality was the same version that was archived for the JOSS paper. Anyways. I am looking into whether there's a way to tinker with what commits get bundled up in the release (so I could bump new functionality to a later version), but I don't yet have it sorted (will update when i do)
