### 2026-04-23 764ab0e
docs: apply JOSS reviewer fixes from PR #156 (links & typos)
Cherry-picks the five-file content delta of PR #156 (smasongarrison/master)
onto the v1.4.1 JOSS-submitted tree:
 - R/DRES-data.R
 - man/DRES.Rd
 - paper/paper.html
 - paper/paper.md
 - vignettes/dySEM.Rmd

No code functionality changes; documentation/typo/link corrections only.

Made-with: Cursor
