+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 `dySEM` is an *R* package [@rcoreteamLanguageEnvironmentStatistical2024] that was created to make it easier for users to deploy the widely popular  `lavaan` package [@rosseelLavaanPackageStructural2012] to fit structural equation models (SEMs) using latent variables to data sets with dependent observations [@littleLongitudinalStructuralEquation2013] collected from interdependent dyads (e.g., romantic partners, pairs of friends, parents, a parent and a child, etc.) [@kennyDyadicDataAnalysis2006]. Initially, the package was exclusively designed to facilitate the testing of dyadic invariance [@sakalukDyadicMeasurementInvariance2021a]--the psychometric equivalence of survey responses across members of dyads. However, the package has since been expanded to include functions that streamline the process of specifying, fitting, and reporting on a variety of models for dyadic data that are amenable to specification in SEM with latent variables [e.g., @kimExtendingActorpartnerInterdependence2022a; @sakalukRenewalDyadicStructural2025].
 
-Currently, `dySEM` faciltates the deployment of several kinds of dyadic SEMs, including:
+Currently, `dySEM` facilitates the deployment of several kinds of dyadic SEMs, including:
 
 - "uni-construct" dyadic models: models in which both members of the dyad are measured on the same, singular construct (e.g., both partners complete a measure of relationship satisfaction)[@sakalukRenewalDyadicStructural2025; @sakalukIntroducingBifactorDyadic2021];
 - "bi-construct" dyadic models: models in which each member of the dyad is measured on two different constructs, with one being used to predict the other (e.g., the popular Actor-Partner Interdependence Model [APIM], and the less popular Common Fate Model [CFM])[@kennyModelsNonIndependenceDyadic1996; @kennyReflectionsActorpartnerInterdependence2018; @kennyDyadicDataAnalysis2006; @kimExtendingActorpartnerInterdependence2022a; @ledermannCommonFateModel2012]
@@ -43,7 +43,7 @@ Though infrequently adopted [@ledermannAnalyzingDyadicData2017; @sakalukRenewalD
 
 But whereas SEM software--like `lavaan` [@rosseelLavaanPackageStructural2012]--can be used to fit dyadic SEMs with latent variables, the process of specifying, fitting, interpreting, and reporting on such models can be arduous and error-prone, particularly for researchers who are new to SEM or to dyadic data analysis. We see these as considerable barriers to adoption of dyadic SEM, and believe that open-source software can help to lower these barriers [@sakalukRenewalDyadicStructural2025].
 
-`dySEM` is available both on GitHub (https://github.com/jsakaluk/dySEM, https://jsakaluk.github.io/dySEM/index.html) and CRAN (https://cran.r-project.org/web/packages/dySEM/index.html), and was created to simplify the process of using dyadic SEM, making it easier for researchers to leverage the power of dyadic SEMs with latent variables in their own work. 
+`dySEM` is available both on GitHub ([repo](https://github.com/jsakaluk/dySEM); [docs](https://jsakaluk.github.io/dySEM/index.html)) and [CRAN](https://cran.r-project.org/web/packages/dySEM/index.html), and was created to simplify the process of using dyadic SEM, making it easier for researchers to leverage the power of dyadic SEMs with latent variables in their own work. 
 Indeed, `dySEM` can typically take an analytic workflow that would otherwise require dozens--if not hundreds--of lines of syntax to be manually written, and reduce it to but a handful or two of lines of (more readable) syntax.
[…]
