### 2026-04-23 02a64f9
chore(release): bump to v1.4.2 for JOSS archival
Made-with: Cursor
