### 2026-04-23 8246306
docs(NEWS): record v1.4.2 JOSS archival release; flag v2.0.0 as post-review
Adds an explicit v1.4.2 section to NEWS.md so master reflects the full
release chronology, and prepends the v2.0.0 section with a provenance
note clarifying that v2.0.0's new functionality (multi-group,
scriptTwoCross, simulation scripters, GUI, orthogonal constraints, etc.)
was developed after and not included in the JOSS-reviewed v1.4.2 release
(openjournals/joss-reviews#10115).

Made-with: Cursor
