+++ NEWS.md
@@ -1,4 +1,6 @@
-# dySEM 2.0.0
+# dySEM 2.0.0 (in development)
+
+* **Note on scope.** v2.0.0 introduces substantial new functionality developed *after* the v1.4.2 JOSS archival release (see openjournals/joss-reviews#10115). These additions (multi-group support, `scriptTwoCross()`, Monte Carlo simulation scripters, GUI support, `"orthogonal"` constraint tokens, etc.) were *not* part of the JOSS-reviewed release and should not be cited as JOSS-reviewed functionality. The JOSS-reviewed and -archived version is v1.4.2 (tag `v1.4.2`, branch `release/joss-1.4.2`).
 
 * `scriptObsAPIM()` now uses the same structural constraint arguments as `scriptAPIM()` (`constr_dy_x_struct`, `constr_dy_y_struct`, `constr_dy_xy_struct`, `includeMeanStruct`, `est_k`), with defaults aligned to an indistinguishable observed APIM. `equate` and `k` are deprecated in favor of `constr_dy_xy_struct` and `est_k`. Orthogonal dyadic covariances are supported via `"orthogonal"` in the `*_struct` vectors. Saving syntax now requires both `writeTo` and `fileName`.
 
@@ -8,6 +10,10 @@
 * Boutique bi-constructc cross-sectional model specification with scriptTwoCross()
 * Initial scripting support for Monte Carlo simulation (via forthcoming updates to dySim: https://jsakaluk.github.io/dySim/)
 * Initial GUI support 
+
+# dySEM 1.4.2
+
+* JOSS review release. Incorporates the documentation and example-script corrections contributed by @smasongarrison in PR #156: link formatting and typo fixes in `R/DRES-data.R`, `man/DRES.Rd`, `paper/paper.md`, `paper/paper.html`, and `vignettes/dySEM.Rmd`. No user-facing code changes relative to 1.4.1; this is the version archived with a DOI for the JOSS submission (openjournals/joss-reviews#10115). Lives on branch `release/joss-1.4.2` at tag `v1.4.2`.
 
 # dySEM 1.4.1
 
