+++ reference/outputUniConstructComp.md
@@ -230,7 +230,7 @@ outputUniConstructComp(
   writeTo = tempdir(),
   fileName = "uni-construct-dyad-models"
 )
-#> Output stored in: /tmp/RtmpHe0EsD/uni-construct-dyad-models.rtf
+#> Output stored in: /tmp/RtmpYJnrx6/uni-construct-dyad-models.rtf
 
 
   
