+++ reference/outputConstraintTab.md
@@ -29,32 +29,31 @@ outputConstraintTab(
 
 - gtTab:
 
-  A logical input indicating whether to generate the table(s) in
+  A logical input indicating whether to generate the output in
   [`gt::gt()`](https://gt.rstudio.com/reference/gt.html) table object
-  format (`TRUE`). By default (`FALSE`), the table(s) are generated in
+  format (`TRUE`). By default (`FALSE`), the output is generated in
   [`tibble::tibble()`](https://tibble.tidyverse.org/reference/tibble.html)
   format. Users can also apply the `writeTo` argument if they wish to
-  export the [`gt::gt()`](https://gt.rstudio.com/reference/gt.html)
-  table object(s).
+  export the `gt:gt()` table object.
 
 - writeTo:
 
-  A character string specifying a directory path to where the output
-  file(s) should be saved. If set to `"."`, the file(s) will be written
-  to the current working directory. The default is `NULL`, and examples
-  use a temporary directory created by
-  [`tempdir()`](https://rdrr.io/r/base/tempfile.html). When dealing with
-  tabular output, `writeTo` is only relevant if `gtTab = TRUE`.
+  A character string specifying a directory path to where the
+  [`gt::gt()`](https://gt.rstudio.com/reference/gt.html) table object
+  should be saved. If set to ".", the file will be written to the
+  current working directory. The default is `NULL`, and examples use a
+  temporary directory created by
+  [`tempdir()`](https://rdrr.io/r/base/tempfile.html). `writeTo` is only
+  relevant if `gtTab = TRUE`.
 
 - fileName:
 
-  A character string specifying a desired base name for the output file.
-  The default is `NULL`. The specified name will be automatically
-  appended with the appropriate file extension (e.g., `.txt` for
-  `lavaan` scripts, `.rtf` for tabular output when `gtTab = TRUE`, or
-  other extensions as appropriate for the output type). If a file with
-  the same name already exists in the user's chosen directory, it will
-  be overwritten.
+  A character string specifying a desired base name for the output
+  [`gt::gt()`](https://gt.rstudio.com/reference/gt.html) file. If a
+  `fileName` is not provided (i.e., `fileName = NULL`), then a default
+  will be used (i.e., "dySEM_table"). The resulting base name will
+  automatically be appended with a `.rtf` file extension. `fileName` is
+  only relevant if `gtTab = TRUE` and `writeTo` is specified.
 
 ## Value
 
@@ -107,232 +106,10 @@ outputConstraintTab(sat.resids.mod,
   filterSig = FALSE,
   gtTab = TRUE, writeTo = tempdir(), fileName = "dCFA_Residual"
 )
-#> Output stored in: /tmp/RtmpYJnrx6/dCFA_Residual.rtf
[…]
