### 2026-05-19 ea734e0
fix: attempt correction of tabel labeling suggested by @samhforbes
