+++ paper/paper.md
@@ -135,7 +135,7 @@ scriptAPIM(dvn, lvxname = "Sat", lvyname = "Com",
 
 There are several other *R*-based tools for dyadic data analysis available for free use, owing to a research culture that promotes the generous provision of methodological and analytic support [e.g., @campbellEstimatingActorPartner2002; @ledermannToolboxProgramsRestructure2015; @stasGivingDyadicData2018]. `dySEM`, however, is distinctive in both its scope, licensing (and underlying values), and analytic robustness. A comparison of `dySEM` versus competitor offerings for dyadic SEM with latent variables and observed variable path analysis is available below in \ref{tab-dysem}. At a high level, `dySEM` is distinguished by being fully open-source (i.e., all source code is available to interested users, and can be modified and built upon by others following a GNU General Public License) and supported by a large number of (transparent) unit-tests, while offering convenient functionality for scripting, calculating, and outputting for a large variety of dyadic SEM models (of which, it is further distinguished by offering the greatest amount of support for [and variety of] latent variable models). And when compared to all alternatives, `dySEM` appears to be the only offering for which unit testing for quality-control is extensive (and growing) and made transparent to the user. We therefore immodestly believe `dySEM` is—and will continue to be—among the more competitive software solutions for dyadic SEM.  
 
-: Comparison of dyadic SEM software. []{label="tab-dysem"}
+: Comparison of dyadic SEM software. \label{tab-dysem}
 
 | Software<br><br><br> | Free | OS Code | OS Tests | Active | Scripting | Documentation |
 |---------------------------|------|---------|----------|--------|-----------|---------------|
