+++ reference/scriptAPIM.md
@@ -36,9 +36,9 @@ scriptAPIM(
 
 - scaleset:
 
-  input character to specify how to set the scale of the latent
-  variable(s). Default is "FF" (fixed-factor; see Details for
-  rationale), but user can specify "MV" (Marker Variable)
+  Input character to specify how to set the scale of the latent
+  variable(s). Default is `"FF"` (fixed-factor; see Details for
+  rationale), but user can specify `"MV"` (Marker Variable).
 
 - lvxname:
 
@@ -60,8 +60,9 @@ scriptAPIM(
 
   input character vector detailing which structural model parameters to
   constrain across dyad members for latent X. Default is c("variances",
-  "means"), but user can specify any combination of "variances" and
-  "means", or "none".
+  "means"), but user can specify any combination of "variances",
+  "means", and/or "orthogonal" (zero dyadic covariance between partners'
+  latent X), or "none".
 
 - constr_dy_y_meas:
 
@@ -74,9 +75,10 @@ scriptAPIM(
 - constr_dy_y_struct:
 
   input character vector detailing which structural model parameters to
-  constrain across dyad members for latent X. Default is c("variances",
-  "means"), but user can specify any combination of "variances" and
-  "means", or "none".
+  constrain across dyad members for latent Y. Default is c("variances",
+  "means"), but user can specify any combination of "variances",
+  "means", and/or "orthogonal" (zero dyadic covariance between partners'
+  latent Y), or "none".
 
 - constr_dy_xy_struct:
 
@@ -116,17 +118,22 @@ scriptAPIM(
 
 - writeTo:
 
-  A character string specifying a directory path to where a .txt file of
-  the resulting lavaan script should be written. If set to “.”, the .txt
-  file will be written to the current working directory. The default is
-  NULL, and examples use a temporary directory created by tempdir().
+  A character string specifying a directory path to where the output
+  file(s) should be saved. If set to `"."`, the file(s) will be written
+  to the current working directory. The default is `NULL`, and examples
+  use a temporary directory created by
+  [`tempdir()`](https://rdrr.io/r/base/tempfile.html). When dealing with
+  tabular output, `writeTo` is only relevant if `gtTab = TRUE`.
 
 - fileName:
 
-  A character string specifying a desired base name for the .txt output
-  file. The default is NULL. The specified name will be automatically
-  appended with the .txt file extension. If a file with the same name
-  already exists in the user's chosen directory, it will be overwritten.
+  A character string specifying a desired base name for the output file.
+  The default is `NULL`. The specified name will be automatically
+  appended with the appropriate file extension (e.g., `.txt` for
+  `lavaan` scripts, `.rtf` for tabular output when `gtTab = TRUE`, or
+  other extensions as appropriate for the output type). If a file with
+  the same name already exists in the user's chosen directory, it will
+  be overwritten.
 
 ## Value
 
@@ -144,7 +151,9 @@ which this function relies on
 Other bi-construct script-writing functions:
 [`scriptBiDy()`](https://jsakaluk.github.io/dySEM/reference/scriptBiDy.md),
 [`scriptCFM()`](https://jsakaluk.github.io/dySEM/reference/scriptCFM.md),
-[`scriptMIM()`](https://jsakaluk.github.io/dySEM/reference/scriptMIM.md)
+[`scriptMIM()`](https://jsakaluk.github.io/dySEM/reference/scriptMIM.md),
+[`scriptObsAPIM()`](https://jsakaluk.github.io/dySEM/reference/scriptObsAPIM.md),
+[`scriptTwoCross()`](https://jsakaluk.github.io/dySEM/reference/scriptTwoCross.md)
 
 ## Examples
 
