+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/L-APIM: L-APIM.html
   articles/M-CDFM: M-CDFM.html
   varnames: varnames.html
-last_built: 2026-05-19T13:24Z
+last_built: 2026-05-19T13:41Z
 urls:
   reference: https://jsakaluk.github.io/dySEM/reference
   article: https://jsakaluk.github.io/dySEM/articles
