### 2026-05-19 396b48e
docs: add JOSS citation (CITATION, CITATION.cff, startup message)
Primary citation is now the JOSS software paper; the 2021 Personal
Relationships article remains as related methodological work in inst/CITATION.

Co-authored-by: Cursor <cursoragent@cursor.com>
