+++ reference/outputUniConstructComp.md
@@ -230,7 +230,7 @@ outputUniConstructComp(
   writeTo = tempdir(),
   fileName = "uni-construct-dyad-models"
 )
-#> Output stored in: /tmp/RtmpyvWYyP/uni-construct-dyad-models.rtf
+#> Output stored in: /tmp/RtmpLNlfMR/uni-construct-dyad-models.rtf
 
 
   
