+++ authors.md
@@ -21,6 +21,22 @@
 Source:
 [`inst/CITATION`](https://github.com/jsakaluk/dySEM/blob/master/inst/CITATION)
 
+Sakaluk, J. K., & Camanto, O. J. (2026). dySEM: An R Package for Dyadic
+Structural Equation Modeling with Latent Variables. Journal of Open
+Source Software, 11(121), 10115. https://doi.org/10.21105/joss.10115
+
+    @Article{,
+      title = {dySEM: An R Package for Dyadic Structural Equation Modeling with Latent Variables},
+      author = {John Kitchener Sakaluk and Omar J. Camanto},
+      journal = {Journal of Open Source Software},
+      year = {2026},
+      volume = {11},
+      number = {121},
+      pages = {10115},
+      doi = {10.21105/joss.10115},
+      url = {https://joss.theoj.org/papers/10.21105/joss.10115},
+    }
+
 Sakaluk, J. K., Fisher, A. N., & Kilshaw, R. E. (2021). Dyadic
 Measurement invariance and its importance for replicability in romantic
 relationship science. Personal Relationships, 28(1), 190-226.
