+++ articles/varnames.md
@@ -8,11 +8,11 @@ library(dySEM)
 #>  
 #>  Please support your academic developers by citing dySEM; this helps us justify to our home institutions the time we spend on expanding and improving dySEM:
 #>  
-#>  Sakaluk, J. K., Fisher, A. N., & Kilshaw, R. K. (2021)
+#>  Sakaluk, J. K., & Camanto, O. J. (2026)
 #> 
-#>                Personal Relationships, 28(1), 190-226. 
+#>                Journal of Open Source Software, 11(121), 10115. 
 #> 
-#>                https://doi.org/10.1111/pere.12341
+#>                https://doi.org/10.21105/joss.10115
 #>  
 #>  dySEM is also heavily dependent on the lavaan package (please cite it too):
 #>  
