### 2026-05-25 02d8715
feat: improved launchScripter docs: post-JOSS acceptance updates to pkgdown and DESCRIPTION
